Fix minor bugs, rename Alien to CleanBot
