+++ source/objects/CleanBot.js
