+++ source/objects/index.js
@@ -1,4 +1,4 @@
 module.exports = {
   Player: require("./Player"),
-  Alien: require("./Alien"),
+  CleanBot: require("./CleanBot"),
 }
